Python/web/oauth_app/requirements.txt
Authlib==1.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.4
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.40
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
